refactor: split file utils
