+++ src/mutahunter/core/utils.py
@@ -1,0 +1,57 @@
+import os
+import shutil
+from mutahunter.core.logger import logger
+
+
+class FileUtils:
+    @staticmethod
+    def read_file(path: str) -> str:
+        try:
+            with open(path, "r") as file:
+                return file.read()
+        except FileNotFoundError:
+            logger.info(f"File not found: {path}")
+        except Exception as e:
+            logger.info(f"Error reading file {path}: {e}")
+            raise
+
+    @staticmethod
+    def backup_code(file_path: str) -> None:
+        backup_path = f"{file_path}.bak"
+        try:
+            shutil.copyfile(file_path, backup_path)
+        except Exception as e:
+            logger.info(f"Failed to create backup file for {file_path}: {e}")
+            raise
+
+    @staticmethod
+    def insert_code(file_path: str, code: str, position: int) -> None:
+        try:
+            with open(file_path, "r") as file:
+                lines = file.read().splitlines()
+            if position == -1:
+                position = len(lines)
+            lines.insert(position, code)
+            with open(file_path, "w") as file:
+                file.write("\n".join(lines))
+
+            # import uuid
+
+            # random_name = str(uuid.uuid4())[:4]
+            # with open(f"{random_name}.java", "w") as file:
+            #     file.write("\n".join(lines))
+        except Exception as e:
+            raise
+
+    @staticmethod
+    def revert(file_path: str) -> None:
+        backup_path = f"{file_path}.bak"
+        try:
+            if os.path.exists(backup_path):
+                shutil.copyfile(backup_path, file_path)
+            else:
+                logger.info(f"No backup file found for {file_path}")
+                raise FileNotFoundError(f"No backup file found for {file_path}")
+        except Exception as e:
+            logger.info(f"Failed to revert file {file_path}: {e}")
+            raise
